+++ src/App.js
@@ -2,10 +2,10 @@ import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 import { ApolloClient, ApolloProvider, InMemoryCache } from '@apollo/client';
 import { Authenticated, Visitor } from './Layouts';
-import RegisterContainer from './Components/pages/Auth/RegisterContainer';
-import LoginContainer from './Components/pages/Auth/LoginContainer';
-import ArticleContainer from './Components/pages/Forum/ArticleContainer';
-import ArticleList from './Components/pages/Forum/ArticleList';
+import RegisterContainer from './Components/Pages/Register/RegisterContainer';
+import LoginContainer from './Components/Pages/Login/LoginContainer';
+import TopicContainer from './Components/Pages/Forum/Topic/TopicContainer';
+import Forum from './Components/Pages/Forum/Forum';
 import './App.css';
 
 const client = new ApolloClient({
@@ -17,12 +17,12 @@ function App() {
   return (
     <ApolloProvider client={client}>
       <Switch>
-        <RouteWrapper exact path='/' component={ArticleList} layout={Visitor} />
+        <RouteWrapper exact path='/' component={Forum} layout={Visitor} />
         <RouteWrapper exact path='/home' component={RegisterContainer} layout={Authenticated} text='home' />
         <RouteWrapper exact path='/register' component={RegisterContainer} layout={Visitor} />
         <RouteWrapper exact path='/login' component={LoginContainer} layout={Visitor} />
-        <RouteWrapper exact path='/topics' history component={ArticleList} layout={Authenticated} text='Forum' />
-        <RouteWrapper path='/topics/:id' component={ArticleContainer} layout={Authenticated} />
+        <RouteWrapper exact path='/topics' history component={Forum} layout={Authenticated} text='Forum' />
+        <RouteWrapper path='/topics/:id' component={TopicContainer} layout={Authenticated} />
       </Switch>
     </ApolloProvider>
   );
